Update
GameServer
 - 몬스터 AI 기능 추가
 - 플레이어가 앉기 상태일 때 HP가 초마다 회복되는 휴식 기능 추가
 - 코드 정리

+++ MMOServer/Common/MathUtil.h
@@ -6,11 +6,11 @@
 
 namespace Jay
 {
-	inline double RadianToDegree(double radian)
+	inline float RadianToDegree(float radian)
 	{
 		return (radian * 180) / PI;
 	}
-	inline double DegreeToRadian(double degree)
+	inline float DegreeToRadian(float degree)
 	{
 		return (degree * PI) / 180;
 	}
